requirements.txt
Django==2.2.5
wget==3.2
gunicorn==19.9.0
psycopg2-binary==2.8.3
factory_boy==2.12.0
mock==3.0.5